api/requirements.txt
# Local dev deps (Windows native, macOS, Linux — all fine).
# The heavy ML stack (torch, allin1, madmom, natten, librosa, etc.) lives in
# requirements-modal.txt and runs inside a Modal container — not here.

modal
fastapi
uvicorn
pytubefix
yt-dlp
pydub
syncedlyrics
opencv-python
scenedetect
openai
python-dotenv
pydantic>=2
psycopg[binary,pool]
redis
pyyaml
moviepy>=2
boto3
imageio-ffmpeg
